+++ lib/project_types/php/cli.rb
@@ -9,6 +9,7 @@ module PHP
     register_command("PHP::Commands::Tunnel", "tunnel")
     register_command("PHP::Commands::Open", "open")
     register_command("PHP::Commands::Populate", "populate")
+    register_command("PHP::Commands::Deploy", "deploy")
 
     require Project.project_filepath("messages/messages")
     register_messages(PHP::Messages::MESSAGES)
@@ -22,6 +23,7 @@ module PHP
     autoload :Connect, Project.project_filepath("commands/connect")
     autoload :Open, Project.project_filepath("commands/open")
     autoload :Populate, Project.project_filepath("commands/populate")
+    autoload :Deploy, Project.project_filepath("commands/deploy")
   end
 
   # define/autoload project specific Tasks