+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Unreleased
 ------
+* [#1197](https://github.com/Shopify/shopify-app-cli/pull/1197): Add `create` command for PHP app projects.
 
 Version 2.4.0
 ------