+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 Unreleased
 ------
-* [#1197](https://github.com/Shopify/shopify-app-cli/pull/1197): Add `create` command for PHP app projects.
-* [#1241](https://github.com/Shopify/shopify-app-cli/pull/1241): Add `serve` command for PHP app projects.
+* [#1197](https://github.com/Shopify/shopify-cli/pull/1197): Add `create` command for PHP app projects.
+* [#1241](https://github.com/Shopify/shopify-cli/pull/1241): Add `serve` command for PHP app projects.
+* [#1243](https://github.com/Shopify/shopify-cli/pull/1243): Add `tunnel` command for PHP app projects.
 
 Version 2.4.0
 ------